Move files into `core`
