+++ src/core/mod.rs
@@ -1,0 +1,3 @@
+pub mod ident;
+pub mod typ;
+pub mod value;
