+++ src/lib.rs
@@ -1,6 +1,4 @@
 #[macro_use]
 extern crate eko_gc_derive;
 
-pub mod ident;
-pub mod typ;
-pub mod value;
+pub mod core;
